+++ General/ApplicationSettings.vb
@@ -65,7 +65,8 @@ Public Class ApplicationSettings
     Public ThumbnailBackgroundColor As Color = Color.AliceBlue
     Public ToolStripRenderModeEx As ToolStripRenderModeEx = ToolStripRenderModeEx.SystemDefault
     Public UIScaleFactor As Single = 1
-    Public UseVapourSynthPortable As Boolean
+    Public UsePortableAviSynth As Boolean = True
+    Public UsePortableVapourSynth As Boolean = True
     Public VapourSynthFilterPreferences As StringPairList
     Public VapourSynthProfiles As List(Of FilterCategory)
     Public Versions As Dictionary(Of String, Integer)